move App layout files into pages
move .App css inline
