--- src/App.css
@@ -1,7 +1,0 @@
-#root,
-.App {
-  /* text-align: center; */
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-}
